+++ bee-transaction/src/atomic/payload/signed_transaction/unlock/signature/ed25519.rs
@@ -1,0 +1,18 @@
+// Copyright 2020 IOTA Stiftung
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+// the License. You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+// an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and limitations under the License.
+
+use alloc::vec::Vec;
+
+#[derive(Debug, Eq, PartialEq)]
+pub struct Ed25519Signature {
+    pub public_key: [u8; 32],
+    pub signature: Vec<u8>,
+}
