+++ bee-transaction/src/atomic/payload/signed_transaction/unlock/signature/mod.rs
@@ -1,0 +1,34 @@
+// Copyright 2020 IOTA Stiftung
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+// the License. You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+// an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and limitations under the License.
+
+mod ed25519;
+mod wots;
+
+pub use ed25519::Ed25519Signature;
+pub use wots::WotsSignature;
+
+#[derive(Debug, Eq, PartialEq)]
+pub enum SignatureUnlock {
+    Wots(WotsSignature),
+    Ed25519(Ed25519Signature),
+}
+
+impl From<WotsSignature> for SignatureUnlock {
+    fn from(signature: WotsSignature) -> Self {
+        Self::Wots(signature)
+    }
+}
+
+impl From<Ed25519Signature> for SignatureUnlock {
+    fn from(signature: Ed25519Signature) -> Self {
+        Self::Ed25519(signature)
+    }
+}
